requirements.txt
dj-database-url==0.4.1
Django==1.9.7
django-bootstrap3==8.1.0
gunicorn==19.6.0
MySQL-python==1.2.4
mysqlclient==1.3.9
olefile==0.44
pbr==1.10.0
Pillow==4.0.0
psycopg2==2.6.1
requests==2.9.1
six==1.10.0
stevedore==1.20.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
whitenoise==2.0.6
djangorestframework==3.6.2